Update Node version in Travis?
I can't figure out why this broke -- according to the Travis docs,
the previous-specified "node" should pull in the latest stable
Node.js release.
But it broke on the spread syntax ({...obj}) for copying objects,
so I'm going to try to update this to the latest LTS Node.js release.
According to the release schedule
[here](https://github.com/nodejs/Release), 10.x is LTS as of 2019-10-30.